tubearchivist/requirements-dev.txt
-r requirements.txt
black
codespell
flake8
isort
pylint
pylint-django
pytest
pytest-django
types-requests
